pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "komitto"
version = "0.1.2"
description = "A CLI tool that generates semantic Japanese commit message prompts for LLMs from Git diff information."
readme = "README.md"
requires-python = ">=3.6"
license = {file = "LICENSE"}
authors = [
  { name = "mxcake" },
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: MIT License",
]
dependencies = [
    "pyperclip",
    "platformdirs",
    "tomli",
    "openai",
    "google-generativeai",
    "anthropic",
]

[project.scripts]
komitto = "komitto.main:main"

[project.urls]
Homepage = "https://github.com/mxcake3893/komitto"

[tool.setuptools.packages.find]
where = ["src"]